Split out Method struct
This commit moves the internal `Method` enum into its own file.
